First implementation of @names
Current implementation links @names in roles to a search for that name, which is a bit more akin to how hashtags work. (Made a hashtag version of this modifier plugin, too.) Would be easy to make it link to the actual maker, just by changing the params on the atnames modifier.